+++ internal/app/wwctl/profile/list/root.go
@@ -1,6 +1,9 @@
 package list
 
-import "github.com/spf13/cobra"
+import (
+	"github.com/spf13/cobra"
+	"github.com/warewulf/warewulf/internal/app/wwctl/completions"
+)
 
 type variables struct {
 	showAll  bool
@@ -18,6 +21,7 @@ func GetCommand() *cobra.Command {
 		Long:                  "This command will display configurations for PROFILE.",
 		RunE:                  CobraRunE(&vars),
 		Aliases:               []string{"ls"},
+		ValidArgsFunction:     completions.Profiles,
 	}
 	baseCmd.PersistentFlags().BoolVarP(&vars.showAll, "all", "a", false, "Show all profile configurations")
 	baseCmd.PersistentFlags().BoolVarP(&vars.showYaml, "yaml", "y", false, "Show profile configurations via yaml format")